replace the list endpoint with new pagination functionality
